requirements.txt
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.2.1
dnspython==2.8.0
fastapi==0.116.1
h11==0.16.0
idna==3.10
outcome==1.3.0.post0
packaging==25.0
pydantic==2.11.7
pydantic_core==2.33.2
pymongo==4.15.1
PySocks==1.7.1
python-dotenv==1.1.1
requests==2.32.4
selenium==4.34.2
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.47.2
trio==0.30.0
trio-websocket==0.12.2
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
